Utilities/pwr_scripts/requirements.txt
# Requires Python 3
matplotlib>=3.8.2
pyserial>=3.5
pyyaml==6.0.1
PyQt5>5.15.11